refactor: extract monitoring cycle logic into separate function
- Updated pnpm workspace dependencies, upgrading 'knip' to version 6.6.1 and adding 'license-checker-rseidelsohn' at version 4.4.2.
- Created a new module 'runMonitoringCycle.ts' to encapsulate the monitoring workflow, including executing Athena queries, processing results, saving CSV, and sending Slack reports.
- Refactored 'main.ts' to utilize the new 'runMonitoringCycle' function, simplifying the main orchestration logic.
