+++ .github/workflows/ci.yml
@@ -7,6 +7,8 @@ on:
   pull_request:
     branches: [main, develop]
 
+  merge_group:
+
 permissions:
   contents: read
   checks: write
@@ -51,34 +53,58 @@ jobs:
     uses: ./.github/workflows/ci-job.yml
     with:
       command: build
+      timeout-minutes: 25
+      upload-artifact-name: workspace-build
+      upload-artifact-path: |
+        packages/**/dist
+        packages/**/*.tsbuildinfo
+        scripts/**/dist
+        scripts/**/*.tsbuildinfo
 
   test:
     name: Test
+    needs: build
     uses: ./.github/workflows/ci-job.yml
     with:
-      pre-command: build
       command: test
+      timeout-minutes: 25
+      download-artifact-name: workspace-build
 
   coverage:
     name: Coverage
+    needs: build
     uses: ./.github/workflows/ci-job.yml
     with:
-      pre-command: build
       command: test:coverage
+      timeout-minutes: 25
+      download-artifact-name: workspace-build
+      upload-artifact-name: coverage-reports
+      upload-artifact-path: |
+        packages/**/coverage/lcov.info
+        scripts/**/coverage/lcov.info
       coverage-lines-threshold: '90'
       coverage-branches-threshold: '80'
       coverage-functions-threshold: '90'
 
+  macos-smoke:
+    name: macOS Smoke
+    uses: ./.github/workflows/ci-job.yml
+    with:
+      runner: macos-latest
+      timeout-minutes: 25
+      command: build:common
+
   # Summary job for branch protection
   ci-success:
     name: CI Success
-    needs: [typecheck, lint, format, knip, scaffold, build, test, coverage]
+    needs: [typecheck, lint, format, knip, scaffold, build, test, coverage, macos-smoke]
     runs-on: ubuntu-latest
+    timeout-minutes: 5
     if: always()
     steps:
       - name: Check all jobs passed
         run: |
-          results=("${{ needs.typecheck.result }}" "${{ needs.lint.result }}" "${{ needs.format.result }}" "${{ needs.knip.result }}" "${{ needs.scaffold.result }}" "${{ needs.build.result }}" "${{ needs.test.result }}" "${{ needs.coverage.result }}")
+          results=("${{ needs.typecheck.result }}" "${{ needs.lint.result }}" "${{ needs.format.result }}" "${{ needs.knip.result }}" "${{ needs.scaffold.result }}" "${{ needs.build.result }}" "${{ needs.test.result }}" "${{ needs.coverage.result }}" "${{ needs.macos-smoke.result }}")
           for result in "${results[@]}"; do
             if [[ "$result" != "success" ]]; then
               echo "One or more checks failed"
