+++ package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "description": "GO Automation Scripts and Tools Monorepo",
   "scripts": {
-    "build": "pnpm build:common && pnpm build:packages && pnpm build:scripts",
-    "build:all": "pnpm build:common && pnpm build:packages && pnpm build:scripts && pnpm build:functions",
+    "build": "pnpm build:packages && pnpm build:scripts",
+    "build:all": "pnpm build:packages && pnpm build:scripts && pnpm build:functions",
     "build:common": "pnpm --filter=@go-automation/go-common build",
     "build:functions": "pnpm -r --filter='./functions/**' build",
     "build:packages": "pnpm -r --filter='./packages/**' build",
@@ -40,6 +40,7 @@
     "lint:fix": "eslint . --fix",
     "lint:packages": "eslint packages",
     "lint:scripts": "eslint scripts",
+    "license:check": "license-checker-rseidelsohn --summary --excludePrivatePackages --onlyAllow \"MIT;Apache-2.0;BSD-2-Clause;BSD-3-Clause;ISC;0BSD;CC0-1.0;Unlicense\"",
     "send:check:ecs:build": "pnpm --filter=send-check-ecs build",
     "send:check:ecs:dev": "pnpm --filter=send-check-ecs dev",
     "send:check:ecs:prod": "pnpm --filter=send-check-ecs start",
@@ -93,6 +94,7 @@
     "eslint-plugin-security": "catalog:",
     "globals": "catalog:",
     "knip": "catalog:",
+    "license-checker-rseidelsohn": "catalog:",
     "prettier": "catalog:",
     "tsx": "catalog:",
     "typescript": "catalog:",
