+++ pnpm-workspace.yaml
@@ -44,7 +44,8 @@ catalog:
   'fuse.js': '^7.3.0'
   'glob': '^13.0.6'
   'globals': '^17.5.0'
-  'knip': '^6.6.0'
+  'knip': '^6.6.1'
+  'license-checker-rseidelsohn': '^4.4.2'
   'luxon': '^3.7.2'
   'prettier': '3.8.3'
   'tsx': '^4.21.0'
